Backend/maja-main/app/pyproject.toml
[tool.poetry]
name = "maja"
version = "0.1.0"
description = ""
authors = ["Kate Yanchenko <kate.yanchenka@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.3"
xrpl-py = "^1.7.0"
djangorestframework = "^3.14.0"
Pillow = "^9.3.0"
django-cors-headers = "^3.13.0"
PyJWT = "^2.6.0"
black = "^22.10.0"
isort = "^5.10.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
